Working on native listener optimizations (reducing JNI Calls): create byte array before the first JNI call to don't call JniTest::toNative() from c++. Test: (1 file, speed = max, ~1350 listener calls, ~255 events per call, lossless)

+++ native/jni-lib/include/api/TimeAndSale.h
@@ -8,7 +8,7 @@ struct TimeAndSale {
   int64_t index;
   int64_t event_flags;
   int32_t time_nano_part;
-  uint16_t exchange_code;
+  int8_t exchange_code;
   double price;
   int64_t size;
   double bid_price;
